Revert "Merge pull request #39 from juice49/features/feed"
This reverts commit b12578f2617acf8a78fbd10739e78b054ba1fc66, reversing
changes made to 1706e7bf3922476c2760738ed8a700adfc8d60f4.
