+++ package.json
@@ -21,7 +21,7 @@
     "groq": "^2.1.4",
     "lodash.upperfirst": "^4.3.1",
     "monstera": "^2.0.0",
-    "next": "^11.1.3-canary.41",
+    "next": "^11.0.0",
     "react": "^17.0.1",
     "react-accessible-headings": "^2.3.0",
     "react-dom": "^17.0.1",
@@ -35,7 +35,6 @@
     "unist-util-visit": "^2.0.3"
   },
   "devDependencies": {
-    "@json-feed-types/1_1": "^1.0.1",
     "@types/node": "^16.4.14",
     "@types/react": "^17.0.0",
     "@types/resize-observer-browser": "^0.1.4",
