+++ src/pages/posts/test.mdx
@@ -1,0 +1,12 @@
+export const meta = {
+  title: 'I am a test post',
+  tags: ['test', 'lorem'],
+  column: 'EverythingIsJavaScript',
+  description: 'Est sapien amet nibh sit, lorem lorem ut elit quis orci nunc.',
+}
+
+# test
+
+- i am
+- a test
+- post
