+++ src/types/post.ts
@@ -10,7 +10,6 @@ export interface PostMeta {
 interface Post {
   meta: PostMeta
   default: string
-  source: string
 }
 
 export default Post
